pyproject.toml
[project]
name = "ebay-watchlist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "gbourdin", email = "german.bourdin@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "humanize>=4.14.0",
    "peewee>=3.18.2",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.6.0",
    "isort>=7.0.0",
    "jupyterlab>=4.4.10",
    "ruff>=0.14.2",
]

[project.scripts]
ebay-watchlist = "ebay_watchlist.cli.main:main"

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"
